+++ src/NodePool/Resurrect/ResurrectInterface.php
@@ -4,20 +4,20 @@
  *
  * @link      https://github.com/elastic/elastic-transport-php
  * @copyright Copyright (c) Elasticsearch B.V (https://www.elastic.co)
- * @license   http://www.apache.org/licenses/LICENSE-2.0 Apache License, Version 2.0
+ * @license   https://opensource.org/licenses/MIT MIT License
  *
  * Licensed to Elasticsearch B.V under one or more agreements.
- * Elasticsearch B.V licenses this file to you under the Apache 2.0 License.
+ * Elasticsearch B.V licenses this file to you under the MIT License.
  * See the LICENSE file in the project root for more information.
  */
 declare(strict_types=1);
 
-namespace Elastic\Transport;
+namespace Elastic\Transport\NodePool\Resurrect;
 
+use Elastic\Transport\NodePool\Node;
 use Psr\Http\Client\ClientInterface;
-use Psr\Http\Message\RequestInterface;
 
-interface ClientAsyncInterface extends ClientInterface
+interface ResurrectInterface
 {
-    public function sendAsync(RequestInterface $request);
+    public function ping(Node $node): bool;
 }