+++ src/Serializer/NDJsonSerializer.php
@@ -4,10 +4,10 @@
  *
  * @link      https://github.com/elastic/elastic-transport-php
  * @copyright Copyright (c) Elasticsearch B.V (https://www.elastic.co)
- * @license   http://www.apache.org/licenses/LICENSE-2.0 Apache License, Version 2.0
+ * @license   https://opensource.org/licenses/MIT MIT License
  *
  * Licensed to Elasticsearch B.V under one or more agreements.
- * Elasticsearch B.V licenses this file to you under the Apache 2.0 License.
+ * Elasticsearch B.V licenses this file to you under the MIT License.
  * See the LICENSE file in the project root for more information.
  */
 declare(strict_types=1);
@@ -50,7 +50,7 @@ class NDJsonSerializer implements SerializerInterface
      * 'type' => (string) specify if the array result should contain object
      *           or array (default is array)
      * 
-     * @return array|ArrayObject
+     * @inheritdoc
      */
     public static function unserialize(string $data, array $options = [])
     {