+++ tests/TransportBuilderTest.php
@@ -4,22 +4,25 @@
  *
  * @link      https://github.com/elastic/elastic-transport-php
  * @copyright Copyright (c) Elasticsearch B.V (https://www.elastic.co)
- * @license   http://www.apache.org/licenses/LICENSE-2.0 Apache License, Version 2.0
+ * @license   https://opensource.org/licenses/MIT MIT License
  *
  * Licensed to Elasticsearch B.V under one or more agreements.
- * Elasticsearch B.V licenses this file to you under the Apache 2.0 License.
+ * Elasticsearch B.V licenses this file to you under the MIT License.
  * See the LICENSE file in the project root for more information.
  */
 declare(strict_types=1);
 
 namespace Elastic\Transport\Test;
 
-use Elastic\Transport\ConnectionPool\SimpleConnectionPool;
 use Elastic\Transport\Exception\CloudIdParseException;
+use Elastic\Transport\NodePool\NodePoolInterface;
 use Elastic\Transport\Transport;
 use Elastic\Transport\TransportBuilder;
+use Http\Discovery\Psr18ClientDiscovery;
+use Http\Discovery\Strategy\MockClientStrategy;
 use Psr\Http\Client\ClientInterface;
 use PHPUnit\Framework\TestCase;
+use Psr\Log\LoggerInterface;
 use Psr\Log\Test\TestLogger;
 
 final class TransportBuilderTest extends TestCase
@@ -32,8 +35,10 @@ final class TransportBuilderTest extends TestCase
     public function setUp(): void
     {
         $this->client = $this->createStub(ClientInterface::class);
-        $this->connectionPool = $this->createStub(SimpleConnectionPool::class);
-        $this->logger = new TestLogger();
+        $this->nodePool = $this->createStub(NodePoolInterface::class);
+        $this->logger = $this->createStub(LoggerInterface::class);
+        
+        Psr18ClientDiscovery::prependStrategy(MockClientStrategy::class);
         $this->builder = TransportBuilder::create();
     }
 
@@ -47,15 +52,26 @@ final class TransportBuilderTest extends TestCase
         $result = $this->builder->setClient($this->client);
 
         $this->assertInstanceOf(TransportBuilder::class, $result);
-        $this->assertEquals($this->builder, $result);
+        $this->assertEquals($this->client, $this->builder->getClient());
     }
 
-    public function testSetConnectionPool()
+    public function testGetClient()
     {
-        $result = $this->builder->setConnectionPool($this->connectionPool);
+        $client = $this->builder->getClient();
+        $this->assertInstanceOf(ClientInterface::class, $client);
+    }
+
+    public function testSetNodePool()
+    {
+        $result = $this->builder->setNodePool($this->nodePool);
 
         $this->assertInstanceOf(TransportBuilder::class, $result);
-        $this->assertEquals($this->builder, $result);
+        $this->assertEquals($this->nodePool, $this->builder->getNodePool());
+    }
+
+    public function testGetNodePool()
+    {
+        $this->assertInstanceOf(NodePoolInterface::class, $this->builder->getNodePool());
     }
 
     public function testSetLogger()
@@ -63,7 +79,12 @@ final class TransportBuilderTest extends TestCase
         $result = $this->builder->setLogger($this->logger);
 
         $this->assertInstanceOf(TransportBuilder::class, $result);
-        $this->assertEquals($this->builder, $result);
+        $this->assertEquals($this->logger, $this->builder->getLogger());
+    }
+
+    public function testGetLogger()
+    {
+        $this->assertInstanceOf(LoggerInterface::class, $this->builder->getLogger());
     }
 
     public function testSetHosts()
@@ -73,6 +94,11 @@ final class TransportBuilderTest extends TestCase
 
         $this->assertInstanceOf(TransportBuilder::class, $result);
         $this->assertEquals($hosts, $this->builder->getHosts());
+    }
+
+    public function testGetHosts()
+    {
+        $this->assertIsArray($this->builder->getHosts());
     }
 
     public function testSetEmptyHosts()
@@ -101,8 +127,25 @@ final class TransportBuilderTest extends TestCase
         $this->assertContains('https://cloud.elastic.co.aaa', $this->builder->getHosts());
     }
 
-    public function testBuildWithDefaultSettings()
+    public function testBuildWithDefault()
     {
         $this->assertInstanceOf(Transport::class, $this->builder->build());
+    }
+
+    public function testBuildWithCustoms()
+    {
+        $this->builder->setClient($this->client);
+        
+        $this->builder->setNodePool($this->nodePool);
+        $this->nodePool->method('setHosts')
+            ->willReturn($this->nodePool);
+
+        $this->builder->setLogger($this->logger);
+
+        $transport = $this->builder->build();
+        $this->assertInstanceOf(Transport::class, $transport);
+        $this->assertEquals($this->client, $transport->getClient());
+        $this->assertEquals($this->nodePool, $transport->getNodePool());
+        $this->assertEquals($this->logger, $transport->getLogger());
     }
 }